+++ ep-sdk/test/specs/services/eventApiProductVersionsService-list.spec.ts
@@ -13,6 +13,9 @@ import {
   EventApiProductResponse,
   EventApiProductsResponse,
   EventApiProduct,
+  EventApiProductVersion,
+  SolaceClassOfServicePolicy,
+  Plan,
 } from '@solace-labs/ep-openapi-node';
 import { 
   EpSdkError,
@@ -23,7 +26,6 @@ import {
   EpSdkEventApiProductAndVersionListResponse,
   EpSdkEventApiProductAndVersionResponse,
   TEpSdkCustomAttributeList,
-  EpSdkBrokerTypes,
 } from '../../../src';
 
 const scriptName: string = path.basename(__filename);
@@ -50,6 +52,20 @@ const ApplicationDomainName = `${TestConfig.getAppId()}/services/${TestSpecId}`;
 let ApplicationDomainId: string | undefined;
 const CustomAttributeList: TEpSdkCustomAttributeList = [
   {
+    name: "PUBLISH_DESTINATION",
+    value: "PublishDestination"
+  },
+  {
+    name: "_X_EP_DEVP_DOMAIN_OWNING_ID_",
+    value: "XEpDevPDomainOwningId"
+  },
+  {
+    name: "_X_EP_DEVP_DOMAIN_SHARING_LIST_",
+    value: "XEpDevPDomainSharingListAttributeValue"
+  },
+];
+const AdditionalCustomAttributeList: TEpSdkCustomAttributeList = [
+  {
     name: "eventApiProduct_1",
     value: "eventApiProduct_1 value"
   },
@@ -58,16 +74,18 @@ const CustomAttributeList: TEpSdkCustomAttributeList = [
     value: "eventApiProduct_2 value"
   }
 ];
-const AdditionalCustomAttributeList: TEpSdkCustomAttributeList = [
-  {
-    name: "eventApiProduct_3",
-    value: "eventApiProduct_3 value"
-  },
-  {
-    name: "eventApiProduct_4",
-    value: "eventApiProduct_4 value"
+
+const EventApiProductVersionPlan_1: Plan = {
+  name: "plan-1",
+  solaceClassOfServicePolicy: {
+    accessType: SolaceClassOfServicePolicy.accessType.EXCLUSIVE,
+    maximumTimeToLive: 1,
+    maxMsgSpoolUsage: 1,
+    messageDeliveryMode: SolaceClassOfServicePolicy.messageDeliveryMode.GUARANTEED,
+    queueType: SolaceClassOfServicePolicy.queueType.COMBINED,
+    type: 'solaceClassOfServicePolicy'
   }
-];
+};
 
 describe(`${scriptName}`, () => {
 
@@ -96,17 +114,7 @@ describe(`${scriptName}`, () => {
   
     after(async() => {
       // delete all application domains
-      // TODO: wait for EP to allow deletion of app domains which still contain event api products
       for(const applicationDomainId of ApplicationDomainIdList) {
-        // delete all event api products first, restriction in EP at the moment
-        const eventApiProductsResponse: EventApiProductsResponse = await EventApiProductsService.getEventApiProducts({
-          applicationDomainId: applicationDomainId
-        });
-        for(const eventApiProduct of eventApiProductsResponse.data) {
-          await EventApiProductsService.deleteEventApiProduct({
-            id: eventApiProduct.id
-          });
-        }
         await EpSdkApplicationDomainsService.deleteById({ applicationDomainId: applicationDomainId });
       }
       await EpSdkApplicationDomainsService.deleteById({ applicationDomainId: ApplicationDomainId });
@@ -136,7 +144,8 @@ describe(`${scriptName}`, () => {
           const x = await EventApiProductsService.createEventApiProductVersion({
             requestBody: {
               eventApiProductId: eventApiProductId,
-              version: EventApiProductVersionString_1
+              version: EventApiProductVersionString_1,
+              plans: [EventApiProductVersionPlan_1]
             }
           });
           // const x = await EventApiProductsService.createEventApiProductVersionForEventApiProduct({
@@ -149,7 +158,7 @@ describe(`${scriptName}`, () => {
           const y = await EventApiProductsService.createEventApiProductVersion({
             requestBody: {
               eventApiProductId: eventApiProductId,
-              version: EventApiProductVersionString_2
+              version: EventApiProductVersionString_2,
             }
           });
           // const y = await EventApiProductsService.createEventApiProductVersionForEventApiProduct({
@@ -267,7 +276,7 @@ describe(`${scriptName}`, () => {
         expect(eventApiProductsResponse.data.length, message).to.equal(NumApplicationDomains);
         for(const eventApiProduct of eventApiProductsResponse.data) {
           // get the latest version for each event api product
-          const latest_EpSdkEventApiProductAndVersionResponse: EpSdkEventApiProductAndVersionResponse = await EpSdkEventApiProductVersionsService.getVersionForEventApiProductId({ 
+          const latest_EpSdkEventApiProductAndVersionResponse: EpSdkEventApiProductAndVersionResponse = await EpSdkEventApiProductVersionsService.getObjectAndVersionForEventApiProductId({ 
             eventApiProductId: eventApiProduct.id,
             stateId: undefined
           });
@@ -277,7 +286,7 @@ describe(`${scriptName}`, () => {
           expect(JSON.stringify(latest_EpSdkEventApiProductAndVersionResponse.meta.versionStringList), message).to.include(EventApiProductVersionString_1);    
           expect(JSON.stringify(latest_EpSdkEventApiProductAndVersionResponse.meta.versionStringList), message).to.include(EventApiProductVersionString_2);    
           // get the version 1 for each event api product
-          const version1_EpSdkEventApiProductAndVersionResponse: EpSdkEventApiProductAndVersionResponse = await EpSdkEventApiProductVersionsService.getVersionForEventApiProductId({ 
+          const version1_EpSdkEventApiProductAndVersionResponse: EpSdkEventApiProductAndVersionResponse = await EpSdkEventApiProductVersionsService.getObjectAndVersionForEventApiProductId({ 
             eventApiProductId: eventApiProduct.id,
             stateId: undefined,
             versionString: EventApiProductVersionString_1
@@ -295,6 +304,74 @@ describe(`${scriptName}`, () => {
       }
     });
 
+    it(`${scriptName}: should get filtered list of versions of for all event api products`, async () => {
+      try {
+        const eventApiProductsResponse: EventApiProductsResponse = await EpSdkEventApiProductsService.listAll({
+          applicationDomainIds: ApplicationDomainIdList,
+          shared: EventApiProductShared,
+        });
+        let message = `eventApiProductsResponse=\n${JSON.stringify(eventApiProductsResponse, null, 2)}`;        
+        expect(eventApiProductsResponse.data, message).to.not.be.undefined;
+        expect(eventApiProductsResponse.data.length, message).to.equal(NumApplicationDomains);
+        for(const eventApiProduct of eventApiProductsResponse.data) {
+          // no filters
+          const eventApiProductVersionList_NoFilters: Array<EventApiProductVersion> = await EpSdkEventApiProductVersionsService.getVersionsForEventApiProductId({
+            eventApiProductId: eventApiProduct.id,
+            stateId: undefined,
+            withAtLeastOnePlan: false,
+            withAtLeastOneAMessagingService: false,
+          });
+          expect(eventApiProductVersionList_NoFilters.length, `length mismatch, eventApiProductVersionList_NoFilters=${JSON.stringify(eventApiProductVersionList_NoFilters, null, 2)}`).to.equal(2);
+          // both filters
+          const eventApiProductVersionList_BothFilters: Array<EventApiProductVersion> = await EpSdkEventApiProductVersionsService.getVersionsForEventApiProductId({
+            eventApiProductId: eventApiProduct.id,
+            stateId: undefined,
+            withAtLeastOnePlan: true,
+            withAtLeastOneAMessagingService: true,
+          });
+          expect(eventApiProductVersionList_BothFilters.length, `length mismatch, eventApiProductVersionList_BothFilters=${JSON.stringify(eventApiProductVersionList_BothFilters, null, 2)}`).to.equal(0);
+        }
+      } catch(e) {
+        if(e instanceof ApiError) expect(false, TestLogger.createApiTestFailMessage('failed')).to.be.true;
+        expect(e instanceof EpSdkError, TestLogger.createNotEpSdkErrorMessage(e)).to.be.true;
+        expect(false, TestLogger.createEpSdkTestFailMessage('failed', e)).to.be.true;
+      }
+    });
+
+    it(`${scriptName}: should filter by at least one plan`, async () => {
+      try {
+        const eventApiProductsResponse: EventApiProductsResponse = await EpSdkEventApiProductsService.listAll({
+          applicationDomainIds: ApplicationDomainIdList,
+          shared: EventApiProductShared,
+        });
+        let message = `eventApiProductsResponse=\n${JSON.stringify(eventApiProductsResponse, null, 2)}`;        
+        expect(eventApiProductsResponse.data, message).to.not.be.undefined;
+        expect(eventApiProductsResponse.data.length, message).to.equal(NumApplicationDomains);
+        for(const eventApiProduct of eventApiProductsResponse.data) {
+          // filter by plan
+          const eventApiProductVersionList_PlanFilter: Array<EventApiProductVersion> = await EpSdkEventApiProductVersionsService.getVersionsForEventApiProductId({
+            eventApiProductId: eventApiProduct.id,
+            stateId: undefined,
+            withAtLeastOnePlan: true,
+            withAtLeastOneAMessagingService: false,
+          });
+          expect(eventApiProductVersionList_PlanFilter.length, `length mismatch, eventApiProductVersionList_PlanFilter=${JSON.stringify(eventApiProductVersionList_PlanFilter, null, 2)}`).to.equal(1);
+          // filter by messaging service
+          const eventApiProductVersionList_MessagingServiceFilter: Array<EventApiProductVersion> = await EpSdkEventApiProductVersionsService.getVersionsForEventApiProductId({
+            eventApiProductId: eventApiProduct.id,
+            stateId: undefined,
+            withAtLeastOnePlan: false,
+            withAtLeastOneAMessagingService: true,
+          });
+          expect(eventApiProductVersionList_MessagingServiceFilter.length, `length mismatch, eventApiProductVersionList_MessagingServiceFilter=${JSON.stringify(eventApiProductVersionList_MessagingServiceFilter, null, 2)}`).to.equal(0);
+        }
+      } catch(e) {
+        if(e instanceof ApiError) expect(false, TestLogger.createApiTestFailMessage('failed')).to.be.true;
+        expect(e instanceof EpSdkError, TestLogger.createNotEpSdkErrorMessage(e)).to.be.true;
+        expect(false, TestLogger.createEpSdkTestFailMessage('failed', e)).to.be.true;
+      }
+    });
+
     it(`${scriptName}: should create event api product for attributes`, async () => {
       try {
         const eventApiProductResponse: EventApiProductResponse = await EventApiProductsService.createEventApiProduct({
@@ -302,6 +379,7 @@ describe(`${scriptName}`, () => {
             applicationDomainId: ApplicationDomainId,
             name: EventApiProductName,
             brokerType: EventApiProduct.brokerType.SOLACE,
+            shared: EventApiProductShared
           }
         });
         EventApiProductId = eventApiProductResponse.data.id;